refactor(web/db): split schema.ts into a per-domain directory
The 802-LOC monolith violated the 350-LOC project rule. Splits
along the existing section banners into eleven files, each under
200 LOC:

  enums.ts          — every pgEnum + the citext custom type
  users.ts          — users + Auth.js standard tables
  content.ts        — submissions, tags, comments, votes, saves
  moderation.ts     — flags + moderation_log
  notifications.ts  — notifications, hidden/muted/email lists, digest_sends
  editorial.ts      — decision_records, override_records, scout_runs
  policy.ts         — policy_decisions
  projects.ts       — projects, project_submissions, project_tags
  metrics.ts        — metrics_daily
  api-tokens.ts     — api_tokens, api_token_usage, api_token_events
  index.ts          — barrel re-exports everything

Forty-one importers of `@/db/schema` resolve through the barrel
unchanged. db/client.ts's `import * as schema from "./schema"`
also resolves transparently — drizzle's namespace import picks up
all named exports from the barrel.

drizzle.config.ts now globs `./src/db/schema/*.ts` so drizzle-kit
sees every file. No migration changes required — the existing
0001..0018 sequence stays canonical; the schema split is purely a
TypeScript organizational change.

No behavioral change. tsc clean, pnpm test green.
