+++ web/drizzle.config.ts
@@ -16,7 +16,7 @@ if (!connectionString) {
 }
 
 export default defineConfig({
-  schema: "./src/db/schema.ts",
+  schema: "./src/db/schema/*.ts",
   out: "./src/db/migrations",
   dialect: "postgresql",
   dbCredentials: { url: connectionString },
